CHANGED baseDir -> directory, removed baseURL
Added README and started formalizing configuration options.

Renamed baseDir to directory, just a nicer name.

The default configuration is to not resolve any files. If you didn’t
configure a directory, then you didn’t mean to grant access to the file
system.

Removed baseURL, this doesn’t seem to be a popular use case, but we
still have urlResolver.