Keep declaration order when inlining into style attribute.
Immutable's Map iteration order is not guaranteed. However, this bug is
only triggered when there are at least 9 declarations to apply. Probably
due to implementation details (I looked briefly but couldn't trace this
to Immutable, so maybe it's Node/V8).

This is especially relevant with shorthand properties. For instance,
`background: red` and `background-color: blue` are different
declarations, but in fact one overrides the other, so order is
important.